feat: implemnet injectable InvestmentService class to share data accros the app
