+++ src/app/app.component.html
@@ -1,3 +1,3 @@
 <app-header />
-<app-user-input (calculateResult)="calculateInvestmentResults($event)" />
-<app-investment-results [investmentResults]="investmentResults()" />
+<app-user-input />
+<app-investment-results />
