+++ src/app/app.component.ts
@@ -1,9 +1,6 @@
 import { Component, signal } from '@angular/core';
 import { HeaderComponent } from './header/header.component';
-import type {
-  InvestmentData,
-  InvestmentResult,
-} from './investment-input.model';
+import type { InvestmentResult } from './investment-input.model';
 import { InvestmentResultsComponent } from './investment-results/investment-results.component';
 import { UserInputComponent } from './user-input/user-input.component';
 
@@ -21,33 +18,4 @@ import { UserInputComponent } from './user-input/user-input.component';
 export class AppComponent {
   investmentResults = signal<InvestmentResult[] | undefined>(undefined);
   // investmentResults?: InvestmentResult[];
-
-  calculateInvestmentResults({
-    initialInvestment,
-    duration,
-    expectedReturn,
-    annualInvestment,
-  }: InvestmentData) {
-    const annualData = [];
-    let investmentValue = initialInvestment;
-
-    for (let i = 0; i < duration; i++) {
-      const year = i + 1;
-      const interestEarnedInYear = investmentValue * (expectedReturn / 100);
-      investmentValue += interestEarnedInYear + annualInvestment;
-      const totalInterest =
-        investmentValue - annualInvestment * year - initialInvestment;
-      annualData.push({
-        year: year,
-        interest: interestEarnedInYear,
-        valueEndOfYear: investmentValue,
-        annualInvestment: annualInvestment,
-        totalInterest: totalInterest,
-        totalAmountInvested: initialInvestment + annualInvestment * year,
-      });
-    }
-
-    // console.log('annualData', annualData);
-    this.investmentResults.set(annualData);
-  }
 }
